Add: added embedding search fucntion to store also cleaned up some code

+++ modules/embeddingsstore/src/main/java/com/theagilemonkeys/ellmental/embeddingsstore/pinecone/QueryVectorResponseSchema.java
@@ -10,7 +10,7 @@ public class QueryVectorResponseSchema {
 
 class Match {
     String id;
-    Integer score;
+    Double score;
     List<Double> values;
     Map<String,String> metadata;
 }
